+++ src/pages/Home/index.jsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components'
-import colors from '../utils/style/colors'
-import { StyledLink } from '../utils/style/Atoms'
-import { useTheme } from '../utils/hooks'
-import HomeIllustration from '../assets/home-illustration.svg'
+import colors from '../../utils/style/colors'
+import { StyledLink } from '../../utils/style/Atoms'
+import { useTheme } from '../../utils/hooks'
+import HomeIllustration from '../../assets/home-illustration.svg'
 
 const HomeWrapper = styled.div`
   display: flex;
