--- src/utils/context/ThemeProvider.jsx
@@ -1,16 +1,0 @@
-import React, { useState, createContext } from 'react'
-
-export const ThemeContext = createContext()
-
-export const ThemeProvider = ({ children }) => {
-  const [theme, setTheme] = useState('light')
-  const toggleTheme = () => {
-    setTheme(theme === 'light' ? 'dark' : 'light')
-  }
-
-  return (
-    <ThemeContext.Provider value={{ theme, toggleTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  )
-}
