+++ src/utils/context/index.jsx
@@ -1,5 +1,20 @@
 import React, { useState, createContext } from 'react'
 
+export const ThemeContext = createContext()
+
+export const ThemeProvider = ({ children }) => {
+  const [theme, setTheme] = useState('light')
+  const toggleTheme = () => {
+    setTheme(theme === 'light' ? 'dark' : 'light')
+  }
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  )
+}
+
 export const SurveyContext = createContext()
 
 export const SurveyProvider = ({ children }) => {
